mushroom_classification_app/pyproject.toml
[tool.poetry]
name = "mushroom-app"
version = "0.1.0"
description = "A Streamlit web app for classifying mushrooms using Pycaret."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
streamlit = "^1.37.1"
pycaret = "^3.3.2"
mlflow = "^2.15.1"
pandas = "<2.2.0"
hydra-core = "^1.3.2"
dvc = "^3.53.1"
kaleido = "0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
